LayoutTests/http/tests/security/cross-frame-access-object-getPrototypeOf-in-put-expected.txt

This tests that you can't get the prototype of the window during [[Put]] operation.

PASS: successfullyParsed should be 'true' and is.

TEST COMPLETE
